First passing test!
Now perhaps, having learned (a little of) the necessary magic,
I can finally start asking more sensible questions of the code.
I've also hacked around quite a lot with the (vry undisciplined)
code to date, but I'm happy to write that all off as a learning
experience.

At present the runs are producing two empty TagSets. So that's
a great place to start looking. Maybe ...