Make FilteredTagSet fit for alpha use.
The FilteredTagSet works exceptionally well, but the Selector
that uses it has bugs that destroy the TagSet's input continuity.
The code I've checked in contains comments indicating my current
theories about the causes, and there are also efficiency gains to
be had (not that I have any performance concerns so far).