Rectify relationships between Selectors and their tagsets.
This has been a journey of discovery, much of it involving
matery of new techniques like writing async methods (and,
more importantly, debugging them).

With the source of the FilteredTagSetSelector class now
down to one line I believe the component APIs are for now
at least usable, although I suspect there are optimisations
that I'll get to as I learn more.