Refactor to move major functionality from App to Screen.
These changes are at test ground for turning the original app,
which was subclassed in multiple programs, each demonstrating
a specific component. Now the app is structured to allow
demonstration screens to be created dynamically, with inputs
provided at run-time.

This minimal version is a proof of concept, having only a
single screen (the TagSet demo) and allows selection of a
(minimally-validated) number of names.

A further innovation is that names are now generated randomly,
currently from a rather limited set. With more than about fifty
names it can start to appear repetitive.