Started work of recasting demos as a single configurable app.
The base screen now takes the additional parameters that are
required to allow the demo_widget method to create the required
TagSet. This is likely far from optimal, but it's unlikely to
matter until the other component types have to be fitted.

Many of the screen's methods have now moved to the app, in
anticipation of more communications being through message
passing.

The app layout is somewhat improved but still needs (a lot of)
work.

Removed a bug in the default formattimg opptions.