Finalise code for initial release.
While the user interface isn't yet how I'd like it to be, the
design seems to be approaching usability and now would be a good
time to get some feedback.

In selector objects the end of selection is now indicated by
hitting the Enter key. New result method returns the selected
elements.